refactor(esbabbler): move slash command parameter mutations into store
